Correct errors given by clang compilers.

+++ src/iga/elm/element_generator.h
@@ -114,8 +114,8 @@ class ElementGenerator {
   }
 
   int Get2DElementNumber(int element_number_xi, int element_number_eta) {
-    util::MultiIndexHandler<2> multi_index_handler(std::array<int, 2>({GetElementList(0).size(),
-                                                                       GetElementList(1).size()}));
+    util::MultiIndexHandler<2> multi_index_handler(std::array<int, 2>({static_cast<int>(GetElementList(0).size()),
+                                                                       static_cast<int>(GetElementList(1).size())}));
     multi_index_handler.SetIndices(std::array<int, 2>({element_number_xi, element_number_eta}));
     return multi_index_handler.Get1DIndex();
   }
